Move 'tilde_expansion()' to the utility module
