+++ src/main/core/support/configuration.rs
@@ -16,6 +16,8 @@ use super::simulation_time::{SIMTIME_INVALID, SIMTIME_ONE_NANOSECOND, SIMTIME_ON
 use super::units::{self, Unit};
 use crate::cshadow as c;
 use crate::host::syscall::format::StraceFmtMode;
+use crate::utility::tilde_expansion;
+
 use log_bindings as c_log;
 
 const START_HELP_TEXT: &str = "\
@@ -1056,26 +1058,6 @@ fn generate_help_strs(
         }
     }
     defaults
-}
-
-pub fn tilde_expansion(path: &str) -> std::path::PathBuf {
-    // if the path begins with a "~"
-    if let Some(x) = path.strip_prefix('~') {
-        // get the tilde-prefix (everything before the first separator)
-        let mut parts = x.splitn(2, '/');
-        let (tilde_prefix, remainder) = (parts.next().unwrap(), parts.next().unwrap_or(""));
-        assert!(parts.next().is_none());
-        // we only support expansion for our own home directory
-        // (nothing between the "~" and the separator)
-        if tilde_prefix.is_empty() {
-            if let Ok(ref home) = std::env::var("HOME") {
-                return [home, remainder].iter().collect::<std::path::PathBuf>();
-            }
-        }
-    }
-
-    // if we don't have a tilde-prefix that we support, just return the unmodified path
-    std::path::PathBuf::from(path)
 }
 
 /// Parses a string as a list of arguments following the shell's parsing rules. This
